requirements.txt
django
djangorestframework
scikit-learn
nltk
python-docx
PyMuPDF
